requirements.txt
Flask
openai
python-dotenv
google-generativeai